moving repofactory to scm module
 - removing the cross module dependency forcing
   the imports to be in particular order
