+++ opengrok-tools/src/main/python/opengrok_tools/scm/__init__.py
@@ -1,10 +1,11 @@
 from .cvs import CVSRepository
 from .git import GitRepository
 from .mercurial import MercurialRepository
+from .repo import RepoRepository
+from .repofactory import get_repository
 from .repository import Repository
 from .svn import SubversionRepository
 from .teamware import TeamwareRepository
-from .repo import RepoRepository
 
 __all__ = [
     'CVSRepository',
@@ -14,4 +15,5 @@ __all__ = [
     'TeamwareRepository',
     'Repository',
     'RepoRepository',
+    'get_repository',
 ]
