+++ opengrok-tools/src/main/python/opengrok_tools/utils/__init__.py
@@ -4,7 +4,6 @@ from . import hook
 from . import opengrok
 from . import parsers
 from . import readconfig
-from . import repofactory
 from . import utils
 from . import webutil
 
@@ -14,7 +13,6 @@ __all__ = [
     'command',
     'commandsequence',
     'hook',
-    'repofactory',
     'webutil',
     'readconfig',
     'parsers',
